requirements.txt
joblib
torch
scikit-learn==0.21.3
pyyaml
shap